refactor: rename envoysubset screener to gwsubset
The screener reads the inference-gateway destination-endpoint-subset
hint from ext_proc dynamic metadata; nothing in the plugin is Envoy
specific. Rename the package, directory, and plugin-type identifier
to gwsubset / gw-subset-screener so the name reflects the protocol it
serves rather than the gateway implementation.

Also tighten the InferenceRequest.Metadata doc comment.
