+++ pkg/epp/framework/interface/scheduling/types.go
@@ -52,8 +52,7 @@ type InferenceRequest struct {
 	Body *fwkrh.InferenceRequestBody
 	// Headers is a map of the request headers.
 	Headers map[string]string
-	// Metadata is the raw Envoy dynamic-metadata map carried by the ext_proc request.
-	// Used by request-control plugins that screen on Envoy-supplied hints (e.g. subset load balancing).
+	// Metadata is the raw dynamic-metadata map carried by the ext_proc request.
 	Metadata map[string]any
 	// Request Objective
 	Objectives RequestObjectives
