+++ pkg/epp/framework/plugins/requestcontrol/screener/gwsubset/plugin_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package envoysubset
+package gwsubset
 
 import (
 	"context"
